# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'Laravel API Documentation'
  description: ''
  version: 1.0.0
servers:
  -
    url: 'https://meningeal-heterogeneously-shalanda.ngrok-free.dev'
...
paths:
...
  '/api/animals/{id}':
    get:
      summary: ''
      operationId: getApiAnimalsId
      description: ''
      parameters: []
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status: false
                  message: 'عفواً، يجب تسجيل الدخول أولاً للوصول لهذا الرابط.'
                  data: null
                properties:
                  status:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'عفواً، يجب تسجيل الدخول أولاً للوصول لهذا الرابط.'
                  data:
                    type: string
                    example: null
                    nullable: true
      tags:
        - Endpoints
      security: []
    delete:
      summary: ''
      operationId: deleteApiAnimalsId
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      security: []
    parameters:
      -
        in: path
        name: id
        description: 'The ID of the animal.'
        example: consequatur
        required: true
        schema:
          type: string
...
  '/api/animals/{animalId}/diagnoses':
    get:
      summary: 'عرض تاريخ التشخيصات لحيوان معين (History)'
      operationId: History
      description: ''
      parameters: []
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status: false
                  message: 'عفواً، يجب تسجيل الدخول أولاً للوصول لهذا الرابط.'
                  data: null
                properties:
                  status:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'عفواً، يجب تسجيل الدخول أولاً للوصول لهذا الرابط.'
                  data:
                    type: string
                    example: null
                    nullable: true
      tags:
        - Endpoints
      security: []
    parameters:
      -
        in: path
        name: animalId
        description: ''
        example: consequatur
        required: true
        schema:
          type: string
...